+++ src/main/java/com/github/mjkuranda/spaceadventure2/states/GameState.java
@@ -55,6 +55,7 @@ public class GameState extends BasicGameState {
         g.drawString(stats.getName(), container.getWidth() - 256, 80);
 
         g.drawString("Remained time [ms]: " + GameData.getRemainingTime(), container.getWidth() - 256,  112);
+        g.drawString("Missile count: " + stats.getMissileCount(), container.getWidth() - 256, 144);
     }
 
     @Override